Moved out radio management functions out of ui_driver No functional change. Build tested with CoIDE, Eclipse, make via Travis.
