+++ mchf-eclipse/drivers/audio/audio_driver.c
@@ -28,7 +28,7 @@
 #include "audio_driver.h"
 #include "audio_management.h"
 #include "dds_table.h"
-#include "ui_driver.h"
+#include "radio_management.h"
 #include "usbd_audio_core.h"
 #include "ui_spectrum.h"
 #include "filters.h"
