+++ mchf-eclipse/drivers/ui/ui.h
@@ -18,17 +18,6 @@
 
 #include "mchf_board.h"
 
-typedef struct BandInfo
-{
-    uint8_t default_pf;
-    uint8_t selector_5W;
-    uint8_t selector_FULL;
-    uint32_t tune;
-    uint32_t size;
-    const char* name;
-} BandInfo;
-
-extern const BandInfo bandInfo[MAX_BAND_NUM];
 
 #endif
